requirement.txt
crispy-bootstrap5==0.6
Django==4.1
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
Pillow==9.2.0
django-filter==22.1
crispy-bootstrap5==0.6
django-extensions==3.2.3
